+++ compiler/lib/compiler/setup/actions/jump.cpp
@@ -12,7 +12,7 @@ int call_action<action::jump>(lua_State* L)
 {
     luaL_checkstring(L, 1);
 
-    lua_createtable(L, 0, 1 + std::size(values));
+    lua_createtable(L, 0, 1 + (int)std::size(values));
     lua_pushliteral(L, "section");
     lua_pushvalue(L, 1);
     lua_settable(L, -3);
