+++ compiler/lib/compiler/setup/options.cpp
@@ -31,7 +31,7 @@ static int create_options_action(lua_State* L)
     lua_pushvalue(L, 1);
     lua_settable(L, -3);
 
-    lua_rawseti(L, -2, length + 1);
+    lua_rawseti(L, -2, (int)length + 1);
 
     lua_pop(L, 2);
